Extract runBunInstall into its own module
Lets ensure-deps.ts (next commit) reuse it without a circular import through src/init/index.ts. Re-exported from src/init/ so the public API is unchanged.
